feat(makefile): add makefile with 'run' target to simplify running the application
The makefile was added to the project with a single target called 'run'. This target allows running the application by executing the 'go run cmd/main.go' command. The target is marked as .PHONY to ensure it always runs, regardless of whether a file with the same name exists. This addition simplifies the process of running the application.